Fix tests for iphone 4s
This includes initializing bitmasks to 0.

+++ objcToxTests/OCTCallTimerTests.m
@@ -63,7 +63,7 @@
 
     [self.callTimer startTimerForCall:call];
 
-    NSInteger delayTime =  dispatch_time(DISPATCH_TIME_NOW, 1.5 * NSEC_PER_SEC);
+    unsigned long long delayTime =  dispatch_time(DISPATCH_TIME_NOW, 1.5 * NSEC_PER_SEC);
 
     dispatch_after(delayTime, dispatch_get_main_queue(), ^{
         if (call.callDuration > 1) {
